+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 - Change NVIDIA open kernel module to be more reliable.
 - Add support for sound on the Steam Deck OLED.
 - Change zram to be more optimized.
+- Change winesapOS log directory from "/etc/winesapos" to "/var/winesapos".
 
 ## 4.0.0 - The Conversion Update